Create html, css and javascript files
Create the layout of the webpage and layout
of elements with combination of flexbox and
grid, add styling and animation to cards

Create the basic logic for the creation
of new entries and removal of said elements,
change of status on the object of said element.

Changes are save on local storage so data is
persistent and won't be overwritten on the reload
of the webpage.

Todo, Mobile responsiveness